Add automated data seed migration.

+++ HoopShoot.Models/Match.cs
@@ -8,7 +8,7 @@
         public int? AwayTeamId { get; set; }
         public Team AwayTeam { get; set; }
 
-        public int HomeTeamScore { get; set; }
-        public int AwayTeamScore { get; set; }
+        public short HomeTeamScore { get; set; }
+        public short AwayTeamScore { get; set; }
     }
 }
